requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.4
cryptography==42.0.7
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.54.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
Markdown==3.6
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python3-openid==3.2.0
requests==2.32.2
requests-oauthlib==2.0.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
